[6261] improve UI Node update speed (#1510)
* only delete nodes when needed

* fix ui customization update

* Kitos data summary extraction script

* add correct order to the script

+++ Core.ApplicationServices/UIConfiguration/UIModuleCustomizationService.cs
@@ -82,7 +82,8 @@ namespace Core.ApplicationServices.UIConfiguration
 
                         var deletedNodes = nodesBefore.Except(nodesAfter).ToList();
 
-                        _repository.DeleteNodes(deletedNodes);
+                        if (deletedNodes.Count > 0)
+                            _repository.DeleteNodes(deletedNodes);
                         _repository.Update(result.Value);
 
                         return Maybe<OperationError>.None;
